requirements.txt
manga_ocr>=0.1.0
Pillow>=8.2.0
easyocr>=1.7.1
PyMuPDF>=1.18.5
zipfile36>=0.1.3
rarfile>=4.0
ebooklib>=0.17.1
numpy>=1.21.0
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0